Fixed a bug where I didn't prepare for chapters which were 'half chapters'... i.e. Naruto has a chapter 594.5. This should be a non-breaking change.

+++ src/MangaEpsilon/Manga/Base/Manga.cs
@@ -58,7 +58,7 @@ namespace MangaEpsilon.Manga.Base
         public Manga ParentManga { get; set; }
         public string ID { get; set; }
         public DateTime ReleaseDate { get; set; }
-        public int VolumeNumber { get; set; }
+        public double VolumeNumber { get; set; }
     }
 
     public class ChapterLight : ChapterBase
